+++ R/huxreg.R
@@ -83,7 +83,7 @@ huxreg <- function (
         bold_signif     = NULL,
         borders         = 0.4,
         outer_borders   = 0.8,
-        note            = "{stars}.",
+        note            = if (is.null(stars)) NULL else "{stars}.",
         statistics      = c("N" = "nobs", "R2" = "r.squared", "logLik", "AIC"),
         coefs           = NULL,
         omit_coefs      = NULL