+++ NEWS.md
@@ -6,6 +6,7 @@ increments reflect backwards-incompatible API changes, not necessarily big chang
 
 * Bugfix: `tidy.tidy_override()` and `glance.tidy_override()` should work even if underlying object
   has no `tidy()` or `glance()` method.
+* Bugfix: huxtables had option clash when `echo = TRUE` in Rmd pdf_document format.
 
 # huxtable 4.3.0