+++ NEWS.md
@@ -7,7 +7,7 @@ version increments reflect backwards-incompatible API changes, not necessarily b
 * Huxtables can now be printed directly in Word documents and Powerpoint presentations,
   thanks to the "flextable" package and recent versions of Pandoc. (Powerpoint printing
   requires Pandoc >= 2.4.0.)
-* New "wrapleft"` and "wrapright" options to `position()` allow text wrapping around tables.
+* New "wrapleft" and "wrapright" options to `position()` allow text wrapping around tables.
 * New `set_outer_border_colors()` and `set_outer_border_styles()` functions, like
   `set_outer_borders()`.
 * Bugfix: `tidy.tidy_override()` and `glance.tidy_override()` should work even if underlying object