+++ NEWS.md
@@ -1,8 +1,9 @@
 
+
 Note that huxtable attempts to follow semantic versioning (https://semver.org). Therefore, major 
 version increments reflect backwards-incompatible API changes, not necessarily big changes.
 
-# huxtable 4.3.0.9000
+# huxtable 4.4.0
 
 * Huxtables can now be printed directly in Word documents and Powerpoint presentations,
   thanks to the `flextable` package and recent versions of Pandoc. (Powerpoint printing